backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
azure-storage-blob==12.19.0
aiofiles==23.2.1
httpx==0.26.0
mediapipe>=0.10.0
opencv-python>=4.6.0
numpy==1.24.4
pyyaml>=6.0
replicate>=0.25.0
